# web-store/src/lib/api/models/Product.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Product = {
    name: string;
    description?: (string | null);
    selling_price: string;
    stock_quantity: number;
    image_url?: (string | null);
    id: number;
    barcode?: (string | null);
    current_price: string;
    cost_price: string;
};