# web-store/src/lib/api/models/ProductCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProductCreate = {
    name: string;
    description?: (string | null);
    selling_price: (number | string);
    cost_price: (number | string);
    stock_quantity?: number;
    barcode?: (string | null);
    image_url?: (string | null);
};