# web-store/src/lib/api/models/ProductUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProductUpdate = {
    name?: (string | null);
    description?: (string | null);
    price?: (number | null);
    stock_quantity?: (number | null);
    barcode?: (string | null);
    image_url?: (string | null);
};